+++ custom_deps/run.sh
@@ -3,19 +3,20 @@ set -e
 
 CONFIG_PATH=/data/options.json
 
-PYPI="$(jq --raw-output '.pypi | join(" ")' $CONFIG_PATH)"
-APK="$(jq --raw-output '.apk | join(" ") // empty' $CONFIG_PATH)"
+PYPI="$(jq --raw-output '.pypi | join(" ")' ${CONFIG_PATH})"
+APK="$(jq --raw-output '.apk | join(" ") // empty' ${CONFIG_PATH})"
 
 # Cleanup old deps
 echo "[Info] Remove old deps"
 rm -rf /config/deps/*
 
 # Need custom apk for build?
-if [ -n "$APK" ]; then
+if [ -n "${APK}" ]; then
     echo "[Info] Install apks for build"
-    if ! ERROR="$(apk add --no-cache "${APK[@]}")"; then
+    # shellcheck disable=SC2086
+    if ! ERROR="$(apk add --no-cache ${APK})"; then
         echo "[Error] Can't install packages!"
-        echo "$ERROR" && exit 1
+        echo "${ERROR}" && exit 1
     fi
 fi
 
@@ -24,9 +25,9 @@ echo "[Info] Install pypi modules into deps"
 export PYTHONUSERBASE=/config/deps
 
 # shellcheck disable=SC2086
-if ! ERROR="$(pip3 install --user --no-cache-dir --prefix= --no-dependencies $PYPI)"; then
-    echo "[Error] Can't install pypi packages!"
-    echo "$ERROR" && exit 1
+if ! ERROR="$(pip3 install --user --no-cache-dir --prefix= --no-dependencies ${PYPI})"; then
+    echo "[Error] Can't install PyPI packages!"
+    echo "${ERROR}" && exit 1
 fi
 
 echo "[Info] done"