+++ custom_deps/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
+## 1
+- Bug fix multiple pip packages v2
+
 ## 0.3
-- Bugfix multible pip packages
+- Bug fix multiple pip packages
 
 ## 0.2
-- Bugfix pip prefix settings
+- Bug fix pip prefix settings
 
 ## 0.1
 - Initial release